Use SparseArrayInterface in DiagonalArrays
