+++ NDTensors/src/lib/DiagonalArrays/examples/README.jl
@@ -2,7 +2,7 @@
 #
 # A Julia `DiagonalArray` type.
 
-using NDTensors.DiagonalArrays: DiagonalArray, DiagIndex, DiagIndices
+using NDTensors.DiagonalArrays: DiagonalArray, DiagIndex, DiagIndices, isdiagindex
 using Test
 
 function main()
@@ -14,7 +14,11 @@ function main()
   d[2, 2, 2] = 22
   @test d[2, 2, 2] == 22
 
-  @test length(d[DiagIndices()]) == 3
+  d_r = reshape(d, 3, 20)
+  @test size(d_r) == (3, 20)
+  @test all(I -> d_r[I] == d[I], LinearIndices(d))
+
+  @test length(d[DiagIndices(:)]) == 3
   @test Array(d) == d
   @test d[DiagIndex(2)] == d[2, 2, 2]
 
@@ -23,15 +27,24 @@ function main()
 
   a = randn(3, 4, 5)
   new_diag = randn(3)
-  a[DiagIndices()] = new_diag
-  d[DiagIndices()] = a[DiagIndices()]
+  a[DiagIndices(:)] = new_diag
+  d[DiagIndices(:)] = a[DiagIndices(:)]
 
-  @test a[DiagIndices()] == new_diag
-  @test d[DiagIndices()] == new_diag
+  @test a[DiagIndices(:)] == new_diag
+  @test d[DiagIndices(:)] == new_diag
 
   permuted_d = permutedims(d, (3, 2, 1))
   @test permuted_d isa DiagonalArray
-  @test permuted_d == d
+  @test permuted_d[DiagIndices(:)] == d[DiagIndices(:)]
+  @test size(d) == (3, 4, 5)
+  @test size(permuted_d) == (5, 4, 3)
+  for I in eachindex(d)
+    if !isdiagindex(d, I)
+      @test iszero(d[I])
+    else
+      @test !iszero(d[I])
+    end
+  end
 
   mapped_d = map(x -> 2x, d)
   @test mapped_d isa DiagonalArray
