+++ NDTensors/src/lib/DiagonalArrays/src/defaults.jl
@@ -1,0 +1,3 @@
+using Compat: Returns
+
+default_size(diag::AbstractVector, n) = ntuple(Returns(length(diag)), n)
