+++ NDTensors/src/lib/DiagonalArrays/src/diagonalmatrix.jl
@@ -1,0 +1,5 @@
+const DiagonalMatrix{T,Diag,Zero} = DiagonalArray{T,2,Diag,Zero}
+
+function DiagonalMatrix(diag::AbstractVector)
+  return DiagonalArray{<:Any,2}(diag)
+end
