+++ NDTensors/src/lib/DiagonalArrays/src/diagonalvector.jl
@@ -1,0 +1,5 @@
+const DiagonalVector{T,Diag,Zero} = DiagonalArray{T,1,Diag,Zero}
+
+function DiagonalVector(diag::AbstractVector)
+  return DiagonalArray{<:Any,1}(diag)
+end
