+++ NDTensors/src/lib/SparseArrayInterface/test/SparseArrayInterfaceTestUtils/SparseArrayInterfaceTestUtils.jl
@@ -1,0 +1,4 @@
+module SparseArrayInterfaceTestUtils
+include("SparseArrays.jl")
+include("DiagonalArrays.jl")
+end
