+++ NDTensors/src/lib/SparseArrayInterface/test/runtests.jl
@@ -1,5 +1,9 @@
 @eval module $(gensym())
 using Compat: Returns, allequal
+using LinearAlgebra: norm
+include("SparseArrayInterfaceTestUtils/SparseArrayInterfaceTestUtils.jl")
+using .SparseArrayInterfaceTestUtils.DiagonalArrays: DiagonalArray
+using .SparseArrayInterfaceTestUtils.SparseArrays: SparseArray
 using Test: @test, @testset, @test_broken, @test_throws
 
 @testset "SparseArrayInterface (eltype=$elt)" for elt in
@@ -14,104 +18,6 @@ using Test: @test, @testset, @test_broken, @test_throws
       CartesianIndex(1, 2)
   end
   @testset "Custom SparseArray" begin
-    using LinearAlgebra: norm
-    using NDTensors.SparseArrayInterface: SparseArrayInterface
-    struct SparseArray{T,N} <: AbstractArray{T,N}
-      data::Vector{T}
-      dims::Tuple{Vararg{Int,N}}
-      index_to_dataindex::Dict{CartesianIndex{N},Int}
-      dataindex_to_index::Vector{CartesianIndex{N}}
-    end
-    function SparseArray{T,N}(dims::Tuple{Vararg{Int,N}}) where {T,N}
-      return SparseArray{T,N}(
-        T[], dims, Dict{CartesianIndex{N},Int}(), Vector{CartesianIndex{N}}()
-      )
-    end
-    SparseArray{T,N}(dims::Vararg{Int,N}) where {T,N} = SparseArray{T,N}(dims)
-    SparseArray{T}(dims::Tuple{Vararg{Int}}) where {T} = SparseArray{T,length(dims)}(dims)
-    SparseArray{T}(dims::Vararg{Int}) where {T} = SparseArray{T}(dims)
-
-    # AbstractArray interface
-    Base.size(a::SparseArray) = a.dims
-    function Base.getindex(a::SparseArray, I...)
-      return SparseArrayInterface.sparse_getindex(a, I...)
-    end
-    function Base.setindex!(a::SparseArray, I...)
-      return SparseArrayInterface.sparse_setindex!(a, I...)
-    end
-    function Base.similar(a::SparseArray, elt::Type, dims::Tuple{Vararg{Int}})
-      return SparseArray{elt}(dims)
-    end
-
-    # Minimal interface
-    SparseArrayInterface.storage(a::SparseArray) = a.data
-    function SparseArrayInterface.index_to_storage_index(
-      a::SparseArray{<:Any,N}, I::CartesianIndex{N}
-    ) where {N}
-      return get(a.index_to_dataindex, I, nothing)
-    end
-    SparseArrayInterface.storage_index_to_index(a::SparseArray, I) = a.dataindex_to_index[I]
-    function SparseArrayInterface.setindex_notstored!(
-      a::SparseArray{<:Any,N}, value, I::CartesianIndex{N}
-    ) where {N}
-      push!(a.data, value)
-      push!(a.dataindex_to_index, I)
-      a.index_to_dataindex[I] = length(a.data)
-      return a
-    end
-
-    # Empty the storage, helps with efficiency in `map!` to drop
-    # zeros.
-    function SparseArrayInterface.empty_storage!(a::SparseArray)
-      empty!(a.data)
-      empty!(a.index_to_dataindex)
-      return empty!(a.dataindex_to_index)
-    end
-
-    # Broadcasting
-    function Broadcast.BroadcastStyle(arraytype::Type{<:SparseArray})
-      return SparseArrayInterface.SparseArrayStyle{ndims(arraytype)}()
-    end
-
-    # Map
-    function Base.map!(f, dest::AbstractArray, src::SparseArray)
-      SparseArrayInterface.sparse_map!(f, dest, src)
-      return dest
-    end
-    function Base.copy!(dest::AbstractArray, src::SparseArray)
-      SparseArrayInterface.sparse_copy!(dest, src)
-      return dest
-    end
-    function Base.copyto!(dest::AbstractArray, src::SparseArray)
-      SparseArrayInterface.sparse_copyto!(dest, src)
-      return dest
-    end
-    function Base.permutedims!(dest::AbstractArray, src::SparseArray, perm)
-      SparseArrayInterface.sparse_permutedims!(dest, src, perm)
-      return dest
-    end
-
-    # Base
-    function Base.:(==)(a1::SparseArray, a2::SparseArray)
-      return SparseArrayInterface.sparse_isequal(a1, a2)
-    end
-    function Base.reshape(a::SparseArray, dims::Tuple{Vararg{Int}})
-      return SparseArrayInterface.sparse_reshape(a, dims)
-    end
-    function Base.iszero(a::SparseArray)
-      return SparseArrayInterface.sparse_iszero(a)
-    end
-    function Base.isreal(a::SparseArray)
-      return SparseArrayInterface.sparse_isreal(a)
-    end
-    function Base.zero(a::SparseArray)
-      return SparseArrayInterface.sparse_zero(a)
-    end
-    function Base.one(a::SparseArray)
-      return SparseArrayInterface.sparse_one(a)
-    end
-
-    # Test
     a = SparseArray{elt}(2, 3)
     @test size(a) == (2, 3)
     @test axes(a) == (1:2, 1:3)
@@ -154,6 +60,17 @@ using Test: @test, @testset, @test_broken, @test_throws
     for I in eachindex(a)
       if I == CartesianIndex(1, 2)
         @test a[I] == 12
+      else
+        @test iszero(a[I])
+      end
+    end
+
+    a = SparseArray{elt}(2, 3)
+    a[1, 2] = 12
+    a = map(x -> 2x, a)
+    for I in eachindex(a)
+      if I == CartesianIndex(1, 2)
+        @test a[I] == 2 * 12
       else
         @test iszero(a[I])
       end
@@ -354,90 +271,6 @@ using Test: @test, @testset, @test_broken, @test_throws
     @test SparseArrayInterface.nstored(a′) == 2
   end
   @testset "Custom DiagonalArray" begin
-    struct DiagonalArray{T,N} <: AbstractArray{T,N}
-      data::Vector{T}
-      dims::Tuple{Vararg{Int,N}}
-    end
-    function DiagonalArray{T,N}(::UndefInitializer, dims::Tuple{Vararg{Int,N}}) where {T,N}
-      return DiagonalArray{T,N}(Vector{T}(undef, minimum(dims)), dims)
-    end
-    function DiagonalArray{T,N}(::UndefInitializer, dims::Vararg{Int,N}) where {T,N}
-      return DiagonalArray{T,N}(undef, dims)
-    end
-    function DiagonalArray{T}(::UndefInitializer, dims::Tuple{Vararg{Int}}) where {T}
-      return DiagonalArray{T,length(dims)}(undef, dims)
-    end
-    function DiagonalArray{T}(::UndefInitializer, dims::Vararg{Int}) where {T}
-      return DiagonalArray{T}(undef, dims)
-    end
-
-    # AbstractArray interface
-    Base.size(a::DiagonalArray) = a.dims
-    function Base.getindex(a::DiagonalArray, I...)
-      return SparseArrayInterface.sparse_getindex(a, I...)
-    end
-    function Base.setindex!(a::DiagonalArray, I...)
-      return SparseArrayInterface.sparse_setindex!(a, I...)
-    end
-    function Base.similar(a::DiagonalArray, elt::Type, dims::Tuple{Vararg{Int}})
-      return DiagonalArray{elt}(undef, dims)
-    end
-
-    # Minimal interface
-    SparseArrayInterface.storage(a::DiagonalArray) = a.data
-    function SparseArrayInterface.index_to_storage_index(
-      a::DiagonalArray{<:Any,N}, I::CartesianIndex{N}
-    ) where {N}
-      !allequal(Tuple(I)) && return nothing
-      return first(Tuple(I))
-    end
-    function SparseArrayInterface.storage_index_to_index(a::DiagonalArray, I)
-      return CartesianIndex(ntuple(Returns(I), ndims(a)))
-    end
-
-    # Broadcasting
-    function Broadcast.BroadcastStyle(arraytype::Type{<:DiagonalArray})
-      return SparseArrayInterface.SparseArrayStyle{ndims(arraytype)}()
-    end
-
-    # Base
-    function Base.iszero(a::DiagonalArray)
-      return SparseArrayInterface.sparse_iszero(a)
-    end
-    function Base.isreal(a::DiagonalArray)
-      return SparseArrayInterface.sparse_isreal(a)
-    end
-    function Base.zero(a::DiagonalArray)
-      return SparseArrayInterface.sparse_zero(a)
-    end
-    function Base.one(a::DiagonalArray)
-      return SparseArrayInterface.sparse_one(a)
-    end
-    function Base.:(==)(a1::DiagonalArray, a2::DiagonalArray)
-      return SparseArrayInterface.sparse_isequal(a1, a2)
-    end
-    function Base.reshape(a::DiagonalArray, dims::Tuple{Vararg{Int}})
-      return SparseArrayInterface.sparse_reshape(a, dims)
-    end
-
-    # Map
-    function Base.map!(f, dest::AbstractArray, src::DiagonalArray)
-      SparseArrayInterface.sparse_map!(f, dest, src)
-      return dest
-    end
-    function Base.copy!(dest::AbstractArray, src::DiagonalArray)
-      SparseArrayInterface.sparse_copy!(dest, src)
-      return dest
-    end
-    function Base.copyto!(dest::AbstractArray, src::DiagonalArray)
-      SparseArrayInterface.sparse_copyto!(dest, src)
-      return dest
-    end
-    function Base.permutedims!(dest::AbstractArray, src::DiagonalArray, perm)
-      SparseArrayInterface.sparse_permutedims!(dest, src, perm)
-      return dest
-    end
-
     # TODO: Test `fill!`.
 
     # Test
@@ -479,10 +312,21 @@ using Test: @test, @testset, @test_broken, @test_throws
 
     a = DiagonalArray(elt[1, 2, 3], (3, 3))
     a_r = reshape(a, 9)
+    @test a_r isa DiagonalArray{elt,1}
+    for I in LinearIndices(a)
+      @test a[I] == a_r[I]
+    end
 
+    # This needs `Base.reshape` with a custom destination
+    # calling `SparseArrayInterface.sparse_reshape!`
+    # in order to specify an appropriate output
+    # type to work.
     a = DiagonalArray(elt[1, 2], (2, 2, 2))
-    # TODO: Make `reshape` for `DiagonalArray` return `SparseArray`.
-    @test_broken reshape(a, 2, 4)
+    a_r = reshape(a, 2, 4)
+    @test a_r isa Matrix{elt}
+    for I in LinearIndices(a)
+      @test a[I] == a_r[I]
+    end
   end
 end
 end
